p108.cpp

+++ p70.cpp
@@ -1,9 +1,8 @@
 #include <bits/stdc++.h>
-
 using namespace std;
 
 int main() {
-    long long n;
+    int n;
     cin >> n;
 
     if (n <= 2) {
@@ -11,8 +10,8 @@ int main() {
         return 0;
     }
 
-    long a = 1, b = 2;
-    for (long i = 3; i <= n; i++) {
+    long long a = 1, b = 2;
+    for (int i = 3; i <= n; i++) {
         long long c = a + b;
         a = b;
         b = c;
